new functions: login, registration, search, send messages.

+++ mychatroom/base/views.py
@@ -1,14 +1,52 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
-from .models import Room, Topic
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
-# Create your views here.
-# rooms=[
-#     {'id':1, 'name':'Learn Python'}, 
-#     {'id':2, 'name':'Learn Java'}, 
-#     {'id':3, 'name':'Learn C++'} 
-# ]
+def loginPage(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method=='POST':
+        username=request.POST.get('username')
+        password=request.POST.get('password')
+        try:
+            user=User.objects.get(username=username)
+            user=authenticate(request, username=username, password=password)
+            if user:
+                login(request, user)
+                return redirect('home')
+            else:
+                messages.error(request, 'incorrect password!')
+        except:
+            messages.error(request, 'user not exist!')
+
+    context={'page':page}
+    return render(request, 'base/login_register.html', context)
+
+def registerUser(request):
+    form=UserCreationForm()
+    if request.method=='POST':
+        form=UserCreationForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Passwords do not match or invalid password!')
+    return render(request, 'base/login_register.html', {'form':form})
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q=request.GET.get('q') if request.GET.get('q')!=None else ''
@@ -24,9 +62,18 @@ def home(request):
 
 def room(request, pk):
     room=Room.objects.get(id=pk) 
-    context={'room':room}
+    room_messages=room.message_set.all().order_by('-created')
+    if request.method=='POST':
+        message=Message.objects.create(
+            user=request.user, 
+            room=room, 
+            body=request.POST.get('body')
+        )
+        return redirect('room', pk=room.id)
+    context={'room':room, 'room_messages':room_messages}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form=RoomForm()
     if request.method == 'POST':
@@ -37,9 +84,12 @@ def createRoom(request):
     context={'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+    if request.user!=room.host:
+        return HttpResponse('Access denied.')
     if request.method=='POST':
         form=RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -48,8 +98,11 @@ def updateRoom(request, pk):
     context={'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room=Room.objects.get(id=pk)
+    if request.user!=room.host:
+        return HttpResponse('Access denied.')
     if request.method=="POST":
         room.delete()
         return redirect('home')
